Remove unused files and modules
Replaced the scanline based ppu with the fifo
ppu and adpted the rest of the project to it
