+++ lib_gb/src/ppu/fifo/mod.rs
@@ -1,6 +1,4 @@
-pub mod fifo_ppu;
-pub mod fifo_register_updater;
-mod background_fetcher;
+pub mod background_fetcher;
+pub mod sprite_fetcher;
 mod fetching_state;
 mod fetcher_state_machine;
-mod sprite_fetcher;
