+++ lib_gb/src/ppu/mod.rs
@@ -5,7 +5,4 @@ pub mod colors;
 pub mod ppu_register_updater;
 pub mod fifo;
 pub mod gfx_device;
-mod normal_sprite;
 mod sprite_attribute;
-mod extended_sprite;
-mod sprite;
