+++ lib_gb/src/ppu/ppu_register_updater.rs
@@ -1,20 +1,20 @@
 use crate::utils::bit_masks::*;
-use super::{ gb_ppu::GbPpu, color::*,  colors::*};
+use super::{color::*, colors::*, gb_ppu::GbPpu, gfx_device::GfxDevice};
 
 const WX_OFFSET:u8 = 7;
 
-pub fn handle_lcdcontrol_register( register:u8, ppu:&mut GbPpu){
-    ppu.screen_enable = (register & BIT_7_MASK) != 0;
-    ppu.window_tile_map_address = (register & BIT_6_MASK) != 0;
-    ppu.window_enable = (register & BIT_5_MASK) != 0;
-    ppu.window_tile_background_map_data_address = (register & BIT_4_MASK) != 0;
-    ppu.background_tile_map_address = (register & BIT_3_MASK) != 0;
-    ppu.sprite_extended = (register & BIT_2_MASK) != 0;
-    ppu.sprite_enable = (register & BIT_1_MASK) != 0;
-    ppu.background_enabled = (register & BIT_0_MASK) != 0;
+pub fn handle_lcdcontrol_register<GFX:GfxDevice>( register:u8, ppu:&mut GbPpu<GFX>){
+    if ppu.lcd_control & BIT_7_MASK != 0 && register & BIT_7_MASK == 0{
+        ppu.turn_off();
+    }
+    else if ppu.lcd_control & BIT_7_MASK == 0 && register & BIT_7_MASK != 0{
+        ppu.turn_on();
+    }
+    
+    ppu.lcd_control = register;
 }
 
-pub fn update_stat_register(register:u8, ppu: &mut GbPpu){
+pub fn update_stat_register<GFX:GfxDevice>(register:u8, ppu: &mut GbPpu<GFX>){
     ppu.h_blank_interrupt_request = register & BIT_3_MASK != 0;
     ppu.v_blank_interrupt_request = register & BIT_4_MASK != 0;
     ppu.oam_search_interrupt_request = register & BIT_5_MASK != 0;
@@ -23,12 +23,12 @@ pub fn update_stat_register(register:u8, ppu: &mut GbPpu){
     ppu.stat_register = register & 0b111_1000;
 }
 
-pub fn set_scx(ppu: &mut GbPpu, value:u8){
-    ppu.background_scroll.x = value;
+pub fn set_scx<GFX:GfxDevice>(ppu: &mut GbPpu<GFX>, value:u8){
+    ppu.bg_pos.x = value;
 }
 
-pub fn set_scy(ppu:&mut GbPpu, value:u8){
-    ppu.background_scroll.y = value;
+pub fn set_scy<GFX:GfxDevice>(ppu:&mut GbPpu<GFX>, value:u8){
+    ppu.bg_pos.y = value;
 }
 
 pub fn handle_bg_pallet_register(register:u8, pallet:&mut [Color;4] ){
@@ -55,27 +55,27 @@ fn get_matching_color(number:u8)->Color{
     };
 }
 
-pub fn handle_wy_register(register:u8, ppu:&mut GbPpu){
-    ppu.window_scroll.y = register;
+pub fn handle_wy_register<GFX:GfxDevice>(register:u8, ppu:&mut GbPpu<GFX>){
+    ppu.window_pos.y = register;
 }
 
-pub fn handle_wx_register(register:u8, ppu:&mut GbPpu){
+pub fn handle_wx_register<GFX:GfxDevice>(register:u8, ppu:&mut GbPpu<GFX>){
     if register < WX_OFFSET{
-        ppu.window_scroll.x = 0;
+        ppu.window_pos.x = 0;
     }
     else{
-        ppu.window_scroll.x = register - WX_OFFSET;
+        ppu.window_pos.x = register - WX_OFFSET;
     }
 }
 
-pub fn get_ly(ppu:&GbPpu)->u8{
-    ppu.current_line_drawn
+pub fn get_ly<GFX:GfxDevice>(ppu:&GbPpu<GFX>)->u8{
+    ppu.ly_register
 }
 
-pub fn get_stat(ppu:&GbPpu)->u8{
+pub fn get_stat<GFX:GfxDevice>(ppu:&GbPpu<GFX>)->u8{
     ppu.stat_register
 }
 
-pub fn set_lyc(ppu:&mut GbPpu, value:u8){
+pub fn set_lyc<GFX:GfxDevice>(ppu:&mut GbPpu<GFX>, value:u8){
     ppu.lyc_register = value;
 }
